renamed icon_card.tsx to IconCard.tsx
