+++ rise-dc-app/src/scheduling_components/icon_components/IconCard.tsx
