+++ rise-dc-app/tsconfig.json
@@ -16,6 +16,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src", "src/scheduling_components/icon_components/icon_card.tsx"],
+  "include": ["src", "src/scheduling_components/icon_components/IconCard.tsx"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
